# Conflicts:
#	hotel-biz/src/main/java/com/mk/hotel/roomtype/service/impl/RoomTypeServiceImpl.java
#	hotel-biz/src/main/resources/common.properties
#	hotel-web/src/main/java/com/mk/execution/pushinfo/ExecutorManager.java
#	hotel-web/src/main/java/com/mk/hotel/hotelinfo/controller/HotelController.java
#	pom.xml

+++ hotel-biz/src/main/java/com/mk/hotel/hotelinfo/service/impl/HotelServiceImpl.java
@@ -8,20 +8,35 @@ import com.mk.framework.JsonUtils;
 import com.mk.framework.excepiton.MyException;
 import com.mk.framework.proxy.http.RedisUtil;
 import com.mk.hotel.common.redisbean.PicList;
+import com.mk.hotel.common.utils.OtsInterface;
 import com.mk.hotel.hotelinfo.HotelService;
 import com.mk.hotel.hotelinfo.bean.HotelLandMark;
 import com.mk.hotel.hotelinfo.dto.HotelDto;
 import com.mk.hotel.hotelinfo.enums.HotelCacheEnum;
+import com.mk.hotel.hotelinfo.enums.HotelSourceEnum;
+import com.mk.hotel.hotelinfo.mapper.HotelFanqieMappingMapper;
 import com.mk.hotel.hotelinfo.mapper.HotelMapper;
 import com.mk.hotel.hotelinfo.model.Hotel;
 import com.mk.hotel.hotelinfo.model.HotelExample;
+import com.mk.hotel.hotelinfo.model.HotelFanqieMapping;
+import com.mk.hotel.hotelinfo.model.HotelFanqieMappingExample;
 import com.mk.hotel.remote.amap.AddressInfoRemoteService;
 import com.mk.hotel.remote.amap.json.AddressByLocationResponse;
+import com.mk.hotel.remote.fanqielaile.hotel.FanqielaileRemoteService;
+import com.mk.hotel.remote.fanqielaile.hotel.json.inn.Inn;
+import com.mk.hotel.remote.fanqielaile.hotel.json.inn.InnList;
+import com.mk.hotel.remote.fanqielaile.hotel.json.proxysalelist.PricePatterns;
+import com.mk.hotel.remote.fanqielaile.hotel.json.proxysalelist.ProxyInns;
+import com.mk.hotel.remote.fanqielaile.hotel.json.proxysalelist.SaleList;
+import com.mk.hotel.remote.fanqielaile.hotel.json.roomstatus.RoomDetailList;
+import com.mk.hotel.remote.fanqielaile.hotel.json.roomstatus.RoomList;
+import com.mk.hotel.remote.fanqielaile.hotel.json.roomtype.RoomType;
+import com.mk.hotel.remote.fanqielaile.hotel.json.roomtype.RoomTypeList;
 import com.mk.hotel.remote.pms.hotel.HotelRemoteService;
 import com.mk.hotel.remote.pms.hotel.json.HotelQueryListRequest;
 import com.mk.hotel.remote.pms.hotel.json.HotelQueryListResponse;
 import com.mk.hotel.roomtype.RoomTypeService;
-import com.mk.hotel.roomtype.service.impl.RoomTypeServiceImpl;
+import com.mk.hotel.roomtype.service.impl.FanqielaileRoomTypeProxyService;
 import com.mk.ots.mapper.LandMarkMapper;
 import com.mk.ots.model.LandMark;
 import com.mk.ots.model.LandMarkExample;
@@ -36,6 +51,7 @@ import redis.clients.jedis.Jedis;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 
@@ -47,6 +63,8 @@ public class HotelServiceImpl implements HotelService {
     @Autowired
     private HotelMapper hotelMapper;
 
+    @Autowired
+    private HotelFanqieMappingMapper hotelFanqieMappingMapper;
     @Autowired
     private LandMarkMapper landMarkMapper;
     @Autowired
@@ -60,6 +78,13 @@ public class HotelServiceImpl implements HotelService {
     @Autowired
     private HotelPicServiceImpl hotelPicService;
 
+    @Autowired
+    private FanqielaileHotelProxyService fanqielaileHotelProxyService;
+    @Autowired
+    private FanqielaileRemoteService fanqielaileRemoteService;
+    @Autowired
+    private FanqielaileRoomTypeProxyService fanqielaileRoomTypeProxyService;
+
     private Logger logger = LoggerFactory.getLogger(HotelServiceImpl.class);
 
     private static List<LandMark> allLandMarkList = new ArrayList<LandMark>();
@@ -72,39 +97,39 @@ public class HotelServiceImpl implements HotelService {
         HotelServiceImpl.allLandMarkList = allLandMarkList;
     }
 
-    public void deleteByFangId(Long id) {
+    public void deleteByFangId(Long id, HotelSourceEnum hotelSourceEnum) {
 
-        if (null == id) {
-            throw new MyException("-99", "-99", "id必填");
+        if (null == id || null == hotelSourceEnum) {
+            throw new MyException("-99", "-99", "id,hotelSourceEnum必填");
         }
 
         //
         HotelExample hotelExample = new HotelExample();
-        hotelExample.createCriteria().andFangIdEqualTo(id);
+        hotelExample.createCriteria().andFangIdEqualTo(id).andSourceTypeEqualTo(hotelSourceEnum.getId());
 
-        List<Hotel> hotelList = hotelMapper.selectByExample(hotelExample);
+        List<Hotel> hotelList = hotelMapper.selectByExampleWithBLOBs(hotelExample);
         if (!hotelList.isEmpty()) {
             Hotel hotel = hotelList.get(0);
             hotel.setIsValid("F");
             hotel.setUpdateBy("hotel_system");
             hotel.setUpdateDate(new Date());
 
-            this.hotelMapper.updateByPrimaryKeySelective(hotel);
+            this.hotelMapper.updateByPrimaryKeyWithBLOBs(hotel);
             this.updateRedisHotel(hotel.getId(), hotel, "HotelService.deleteByFangId");
         }
 
     }
 
-    public void saveOrUpdateByFangId(HotelDto hotelDto) {
+    public void saveOrUpdateByFangId(HotelDto hotelDto, HotelSourceEnum hotelSourceEnum) {
 
-        if (null == hotelDto || null == hotelDto.getFangId()) {
-            throw new MyException("-99", "-99", "hotelDto fangId必填");
+        if (null == hotelDto || null == hotelDto.getFangId() || null == hotelSourceEnum) {
+            throw new MyException("-99", "-99", "hotelDto fangId hotelSourceEnum必填");
         }
 
         //
         HotelExample hotelExample = new HotelExample();
-        hotelExample.createCriteria().andFangIdEqualTo(hotelDto.getFangId());
-        List<Hotel> hotelList = hotelMapper.selectByExample(hotelExample);
+        hotelExample.createCriteria().andFangIdEqualTo(hotelDto.getFangId()).andSourceTypeEqualTo(hotelSourceEnum.getId());
+        List<Hotel> hotelList = hotelMapper.selectByExampleWithBLOBs(hotelExample);
 
         //
         Long hotelId = null;
@@ -120,6 +145,7 @@ public class HotelServiceImpl implements HotelService {
             dbHotel.setCreateBy("hotel_system");
             dbHotel.setUpdateDate(new Date());
             dbHotel.setUpdateBy("hotel_system");
+            dbHotel.setSourceType(hotelSourceEnum.getId());
             //
             this.hotelMapper.insert(dbHotel);
             hotelId = dbHotel.getId();
@@ -168,9 +194,10 @@ public class HotelServiceImpl implements HotelService {
 
             hotel.setUpdateDate(new Date());
             hotel.setUpdateBy("hotel_system");
+            hotel.setSourceType(hotelSourceEnum.getId());
 
             hotelId = hotel.getId();
-            this.hotelMapper.updateByPrimaryKeySelective(hotel);
+            this.hotelMapper.updateByPrimaryKeyWithBLOBs(hotel);
             hotelDto.setId(hotelId);
             this.updateRedisHotel(hotelId, hotel, "HotelService.saveOrUpdateByFangId(HotelDto)");
 
@@ -178,7 +205,7 @@ public class HotelServiceImpl implements HotelService {
 
         //roomTypeList
         if (null != hotelDto.getRoomTypeDtoList() && !hotelDto.getRoomTypeDtoList().isEmpty()) {
-            this.roomTypeService.saveOrUpdateByHotelId(hotelId, hotelDto.getRoomTypeDtoList());
+            this.roomTypeService.saveOrUpdateByHotelId(hotelId, hotelDto.getRoomTypeDtoList(), hotelSourceEnum);
         }
     }
 
@@ -186,7 +213,7 @@ public class HotelServiceImpl implements HotelService {
     public List<HotelDto> findHotelByName(String hotelName, String cityCode) {
         HotelExample hotelExample = new HotelExample();
         hotelExample.createCriteria().andNameEqualTo(hotelName.trim()).andCityCodeEqualTo(cityCode.trim());
-        List<Hotel> hotelList = hotelMapper.selectByExample(hotelExample);
+        List<Hotel> hotelList = hotelMapper.selectByExampleWithBLOBs(hotelExample);
         if(CollectionUtils.isEmpty(hotelList)){
             return null;
         }
@@ -214,7 +241,7 @@ public class HotelServiceImpl implements HotelService {
     public HotelDto findByName(String name) {
         HotelExample hotelExample = new HotelExample();
         hotelExample.createCriteria().andNameEqualTo(name);
-        List<Hotel> hotelList = hotelMapper.selectByExample(hotelExample);
+        List<Hotel> hotelList = hotelMapper.selectByExampleWithBLOBs(hotelExample);
         if(CollectionUtils.isEmpty(hotelList)){
             return null;
         }
@@ -223,11 +250,11 @@ public class HotelServiceImpl implements HotelService {
         return dto;
     }
 
-    public HotelDto findByFangId(Long fangId) {
+    public HotelDto findByFangId(Long fangId, HotelSourceEnum hotelSourceEnum) {
 
         HotelExample hotelExample = new HotelExample();
-        hotelExample.createCriteria().andFangIdEqualTo(fangId);
-        List<Hotel> hotelList = hotelMapper.selectByExample(hotelExample);
+        hotelExample.createCriteria().andFangIdEqualTo(fangId).andSourceTypeEqualTo(hotelSourceEnum.getId());
+        List<Hotel> hotelList = hotelMapper.selectByExampleWithBLOBs(hotelExample);
 
         if (hotelList.isEmpty()) {
             return null;
@@ -342,7 +369,7 @@ public class HotelServiceImpl implements HotelService {
             Integer intProvCode = null;
             try {
                 intProvCode = Integer.parseInt(hotel.getProvCode());
-            } catch (NumberFormatException e) {
+            } catch (Exception e) {
                 e.printStackTrace();
                 Cat.logError(e);
             }
@@ -350,7 +377,7 @@ public class HotelServiceImpl implements HotelService {
             Integer intCityCode = null;
             try {
                 intCityCode = Integer.valueOf(hotel.getCityCode());
-            } catch (NumberFormatException e) {
+            } catch (Exception e) {
                 e.printStackTrace();
                 Cat.logError(e);
             }
@@ -358,7 +385,7 @@ public class HotelServiceImpl implements HotelService {
             Integer intDisCode = null;
             try {
                 intDisCode = Integer.valueOf(hotel.getDisCode());
-            } catch (NumberFormatException e) {
+            } catch (Exception e) {
                 e.printStackTrace();
                 Cat.logError(e);
             }
@@ -367,7 +394,7 @@ public class HotelServiceImpl implements HotelService {
             if(StringUtils.isNotBlank(townCode)){
                 try {
                     intTownCode = Integer.parseInt(townCode);
-                } catch (NumberFormatException e) {
+                } catch (Exception e) {
                     e.printStackTrace();
                     Cat.logError(e);
                 }
@@ -424,6 +451,8 @@ public class HotelServiceImpl implements HotelService {
             //
             hotelInRedis.setCacheTime(strDate);
             hotelInRedis.setCacheFrom(cacheFrom);
+
+            hotelInRedis.setHotelSource(hotel.getSourceType());
             //
             jedis.set(hotelKeyName, JsonUtils.toJson(hotelInRedis));
 
@@ -506,5 +535,166 @@ public class HotelServiceImpl implements HotelService {
         return hotelLandMark;
     }
 
+    public List<String> mergeFanqieHotel (){
 
+        List<String> result = new ArrayList<String>();
+        //
+        SaleList saleList = this.fanqielaileRemoteService.queryHotelList();
+
+        if (null == saleList) {
+            return result;
+        }
+
+        //
+        List<ProxyInns> proxyInnsList = saleList.getProxyInns();
+        if (null != proxyInnsList) {
+
+            for (ProxyInns proxyInns: proxyInnsList) {
+                String proxyInnJson = JsonUtils.toJson(proxyInns);
+                result.add(proxyInnJson);
+            }
+        }
+
+        //for debug
+//        result.add("{\n" +
+//                "      \"innId\": 4831,\n" +
+//                "      \"pricePatterns\": [\n" +
+//                "        {\n" +
+//                "          \"pattern\": \"2\",\n" +
+//                "          \"accountId\": 27941\n" +
+//                "        }\n" +
+//                "      ]\n" +
+//                "    }");
+
+        return result;
+    }
+
+    public void mergeFangqieHotelByProxyInnJson (String proxyInnJson) {
+
+        ProxyInns proxyInns = JsonUtils.fromJson(proxyInnJson, ProxyInns.class);
+
+        List<PricePatterns> pricePatternsList = proxyInns.getPricePatterns();
+        Integer innId = proxyInns.getInnId();
+
+        if (null != pricePatternsList) {
+
+            for (PricePatterns pricePatterns : pricePatternsList) {
+
+                //策略模式1:底价模式 2:卖价模式
+                String pattern = pricePatterns.getPattern();
+                Integer intPattern = null;
+                try {
+                    intPattern = Integer.parseInt(pattern);
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+
+                //
+                Integer accountId =  pricePatterns.getAccountId();
+
+                //update
+                this.fanqielaileHotelProxyService.saveOrUpdateMapping(innId.longValue(), intPattern, accountId.longValue());
+
+                /*
+                    1 酒店
+                 */
+                Hotel hotel = null;
+                InnList innList = this.fanqielaileRemoteService.queryInn(accountId.longValue());
+                if (null != innList) {
+                    List<Inn> inns = innList.getList();
+                    for (Inn inn : inns) {
+                        try {
+                            hotel = this.fanqielaileHotelProxyService.saveOrUpdateHotel(innId, inn);
+                        } catch (Exception e) {
+                            Cat.logError(e);
+                        }
+                    }
+                }
+
+                /*
+                    2 房型
+                 */
+                RoomTypeList roomTypeList = this.fanqielaileRemoteService.queryRoomType(
+                        accountId.longValue());
+                if (null != roomTypeList) {
+                    List<RoomType> roomTypes = roomTypeList.getList();
+                    for (RoomType roomType : roomTypes) {
+                        this.fanqielaileRoomTypeProxyService.saveOrUpdateRoomType(innId, hotel.getId(), roomType);
+                    }
+                }
+
+                //
+                OtsInterface.initHotel(hotel.getId());
+            }
+        }
+    }
+
+    public List<String> mergeFangqieRoomStatus () {
+
+        HotelFanqieMappingExample example = new HotelFanqieMappingExample();
+        example.createCriteria().andIsValidEqualTo("T");
+        List<HotelFanqieMapping> hotelFanqieMappingList = this.hotelFanqieMappingMapper.selectByExample(example);
+
+        //
+        List<String> result = new ArrayList<String>();
+        for (HotelFanqieMapping mapping : hotelFanqieMappingList) {
+            result.add(JsonUtils.toJson(mapping));
+        }
+        return result;
+    }
+
+    public void mergeFangqieRoomStatusByHotelFanqieMappingJson (String hotelFanqieMappingJson) {
+        HotelFanqieMapping hotelFanqieMapping = JsonUtils.fromJson(hotelFanqieMappingJson, HotelFanqieMapping.class);
+
+        Long accountId = hotelFanqieMapping.getAccountId();
+        Long innId = hotelFanqieMapping.getInnId();
+
+        //
+        Hotel hotel = null;
+        HotelExample hotelExample = new HotelExample();
+        hotelExample.createCriteria().andFangIdEqualTo(innId).andSourceTypeEqualTo(HotelSourceEnum.FANQIE.getId()).andIsValidEqualTo("T");
+        List<Hotel> hotelList = this.hotelMapper.selectByExample(hotelExample);
+        if (hotelList.isEmpty()) {
+            return;
+        } else {
+            hotel = hotelList.get(0);
+        }
+
+        //查询30天的房态
+        Calendar calendar = Calendar.getInstance();
+        calendar.setTime(new Date());
+        calendar.add(Calendar.DAY_OF_MONTH,30);
+        Date nextMonth = calendar.getTime();
+
+        //
+        RoomList roomList = this.fanqielaileRemoteService.queryRoomStatus(
+                accountId, new Date() , nextMonth);
+        if (null != roomList) {
+
+            List<RoomDetailList> roomDetailLists = roomList.getList();
+            for (RoomDetailList detailList : roomDetailLists) {
+                this.fanqielaileRoomTypeProxyService.saveOrUpdateRoomDetail(hotel.getId(), detailList);
+            }
+        }
+
+        //
+        OtsInterface.initHotel(hotel.getId());
+    }
+
+
+    public HotelFanqieMapping findHotelMappingByHotelId(Long hotelId) {
+        Hotel hotel = hotelMapper.selectByPrimaryKey(hotelId);
+        if(hotel == null) {
+            return null;
+        }
+        HotelFanqieMappingExample example = new HotelFanqieMappingExample();
+        example.createCriteria().andInnIdEqualTo(hotel.getFangId());
+        List<HotelFanqieMapping> hotelFanqieMappingList =  hotelFanqieMappingMapper.selectByExample(example);
+        if(org.apache.commons.collections.CollectionUtils.isEmpty(hotelFanqieMappingList)){
+            return null;
+        }else if(hotelFanqieMappingList.size() > 1){
+            throw new MyException("酒店信息错误,根据hotelId找到多个映射酒店信息");
+        }
+        return hotelFanqieMappingList.get(0);
+    }
 }
